Add mapping of OIDC client user to iRODS user
When OIDC is enabled, switch off Sqyrrl's current mode of showing only
public (iRODS group) data, but showing it to all clients.

Instead, when OIDC is enabled, Sqyrrl maps the authenticated OIDC
client user to an iRODS user of the same name and then uses the
standard iRODS permissions model. In this mode a client can see
"public" data only if their mapped iRODS user is a member of the
public iRODS group.

Sqyrrl's HTTP session manager is now passed to its constructor so that
it ois accessible to be externally configured and also to simplify
testing because fake OIDC sessions can be set up to test the HTTP
handlers, without the need for an OIDC server or mocks.