Fix access for federated users with the same name as a local user
Ensure that when a data obect has "foo#localZone:read" in its ACL, it
allows read operations by "foo#localZone" and "foo" (implied
"localZone"), but not by "foo#remoteZone".

To do this we need to distinguish in the IsReadableByUser function
between the local zone (the zone to which the Sqyrrl server is logged
in) and the home zone of the user being tested. In practice the
local/login zone and the home zone of the iRODS user that wants access
are very likely to be the same. However, conflating the two zones
makes the IsReadableByUser function confusing.